OSS-24 CntEstadoCuentas Add endpoint Create

+++ Convertidor/Data/Entities/Cnt/CNT_ESTADO_CUENTAS.cs
@@ -7,8 +7,8 @@
         public string NUMERO_ESTADO_CUENTA { get; set; } = string.Empty;
         public DateTime FECHA_DESDE { get; set; }
         public DateTime FECHA_HASTA { get; set; }
-        public int SALDO_INICIAL { get; set; }
-        public int SALDO_FINAL { get; set; }
+        public decimal SALDO_INICIAL { get; set; }
+        public decimal SALDO_FINAL { get; set; }
         public string EXTRA1 { get; set; } = string.Empty;
         public string EXTRA2 { get; set; } = string.Empty;
         public string EXTRA3 { get; set; } = string.Empty;
